+++ general/book-assert/index.ts
@@ -1,5 +1,5 @@
 export { Book } from "./Book";
-export { BookIndex } from "./BookIndex";
+export { BookIndexer } from "./BookIndexer";
 export { Chapter } from "./Chapter";
 export { main } from "./main";
 export { run } from "./run";
